bela rework part I

+++ cmake/SetupLibTorch.cmake
@@ -1,5 +1,9 @@
 # torch stopped uploading the binaries for macOS x86_64, but we use self-built binaries
-set(LIBTORCH_VERSION 2.4.1)
+if(UNIX AND NOT APPLE AND CMAKE_SYSTEM_PROCESSOR STREQUAL "armv7a")
+    set(LIBTORCH_VERSION 2.5.1)
+else()
+    set(LIBTORCH_VERSION 2.4.1)
+endif()
 
 if (NOT WIN32)
     set(TORCH_BUILD_TYPE "")
@@ -33,10 +37,14 @@ else()
     if(UNIX AND NOT APPLE)
         if (CMAKE_SYSTEM_PROCESSOR STREQUAL "aarch64")
             set(LIB_LIBTORCH_PRE_BUILD_LIB_NAME "libtorch-${LIBTORCH_VERSION}-Linux-aarch64")
+            set(LIB_LIBTORCH_PRE_BUILD_LIB_TYPE "zip")
         elseif(CMAKE_SYSTEM_PROCESSOR STREQUAL "x86_64")
             set(LIB_LIBTORCH_PRE_BUILD_LIB_NAME "libtorch-${LIBTORCH_VERSION}-Linux-x86_64")
+            set(LIB_LIBTORCH_PRE_BUILD_LIB_TYPE "zip")
+        elseif(CMAKE_SYSTEM_PROCESSOR STREQUAL "armv7a")
+            set(LIB_LIBTORCH_PRE_BUILD_LIB_NAME "pytorch-v${LIBTORCH_VERSION}")
+            set(LIB_LIBTORCH_PRE_BUILD_LIB_TYPE "tar.gz")
         endif()
-        set(LIB_LIBTORCH_PRE_BUILD_LIB_TYPE "zip")
     endif()
 
     if(UNIX AND APPLE)
@@ -51,6 +59,8 @@ else()
     if (NOT DEFINED LIBTORCH_URL)
         if(WIN32)
             set(LIBTORCH_URL https://download.pytorch.org/libtorch/cpu/${LIB_LIBTORCH_PRE_BUILD_LIB_NAME}.${LIB_LIBTORCH_PRE_BUILD_LIB_TYPE})
+        elseif(UNIX AND NOT APPLE AND CMAKE_SYSTEM_PROCESSOR STREQUAL "armv7a")
+            set(LIBTORCH_URL https://github.com/pelinski/bela-torch/releases/download/v${LIBTORCH_VERSION}/${LIB_LIBTORCH_PRE_BUILD_LIB_NAME}.${LIB_LIBTORCH_PRE_BUILD_LIB_TYPE})
         else()
             set(LIBTORCH_URL https://github.com/faressc/libtorch-cpp-lib/releases/download/v${LIBTORCH_VERSION}/${LIB_LIBTORCH_PRE_BUILD_LIB_NAME}.${LIB_LIBTORCH_PRE_BUILD_LIB_TYPE})
         endif()
@@ -63,14 +73,21 @@ else()
     file(DOWNLOAD ${LIBTORCH_URL} ${LIBTORCH_PATH} STATUS LIBTORCH_DOWNLOAD_STATUS SHOW_PROGRESS)
     list(GET LIBTORCH_DOWNLOAD_STATUS 0 LIBTORCH_DOWNLOAD_STATUS_NO)
 
-    file(ARCHIVE_EXTRACT
+    if(UNIX AND NOT APPLE AND CMAKE_SYSTEM_PROCESSOR STREQUAL "armv7a")
+        execute_process(
+            COMMAND tar -xzf ${LIBTORCH_PATH} -C ${CMAKE_CURRENT_SOURCE_DIR}/modules/libtorch-${LIBTORCH_VERSION}${TORCH_BUILD_TYPE}/
+            WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/modules/libtorch-${LIBTORCH_VERSION}${TORCH_BUILD_TYPE}/
+        )
+    else()
+        file(ARCHIVE_EXTRACT
             INPUT ${CMAKE_BINARY_DIR}/import/${LIB_LIBTORCH_PRE_BUILD_LIB_NAME}.${LIB_LIBTORCH_PRE_BUILD_LIB_TYPE}
             DESTINATION ${CMAKE_CURRENT_SOURCE_DIR}/modules/libtorch-${LIBTORCH_VERSION}${TORCH_BUILD_TYPE}/)
+    endif()
 
-    if (EXISTS ${CMAKE_CURRENT_SOURCE_DIR}/modules/libtorch-${LIBTORCH_VERSION}${TORCH_BUILD_TYPE}/libtorch)
+    if(EXISTS ${CMAKE_CURRENT_SOURCE_DIR}/modules/libtorch-${LIBTORCH_VERSION}${TORCH_BUILD_TYPE}/libtorch)
         file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/modules/libtorch-${LIBTORCH_VERSION}${TORCH_BUILD_TYPE}/libtorch/ DESTINATION ${CMAKE_CURRENT_SOURCE_DIR}/modules/libtorch-${LIBTORCH_VERSION}${TORCH_BUILD_TYPE}/)
         file(REMOVE_RECURSE ${CMAKE_CURRENT_SOURCE_DIR}/modules/libtorch-${LIBTORCH_VERSION}${TORCH_BUILD_TYPE}/libtorch/)
-    else()
+    elseif(EXISTS ${CMAKE_CURRENT_SOURCE_DIR}/modules/libtorch-${LIBTORCH_VERSION}${TORCH_BUILD_TYPE}/${LIB_LIBTORCH_PRE_BUILD_LIB_NAME})
         file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/modules/libtorch-${LIBTORCH_VERSION}${TORCH_BUILD_TYPE}/${LIB_LIBTORCH_PRE_BUILD_LIB_NAME}/ DESTINATION ${CMAKE_CURRENT_SOURCE_DIR}/modules/libtorch-${LIBTORCH_VERSION}${TORCH_BUILD_TYPE}/)
         file(REMOVE_RECURSE ${CMAKE_CURRENT_SOURCE_DIR}/modules/libtorch-${LIBTORCH_VERSION}${TORCH_BUILD_TYPE}/${LIB_LIBTORCH_PRE_BUILD_LIB_NAME}/)
     endif()
